Begin conversion to use project layout for hikes instead
